+++ src/main/groovy/com/trevorism/model/AuthorizeRequest.groovy
@@ -1,0 +1,7 @@
+package com.trevorism.model
+
+class AuthorizeRequest {
+    String redirectUri
+    String state
+    String guid
+}