+++ src/main/groovy/com/trevorism/model/Oauth2State.groovy
@@ -1,0 +1,9 @@
+package com.trevorism.model
+
+class Oauth2State {
+    String nonce
+    String returnUrl
+    String guid
+    String redirectUri
+    String state
+}